Backend/Milk_Saas/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cssselect2==0.7.0
Django==5.1.6
django-filter==24.3
django-redis==5.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
fonttools==4.56.0
gunicorn==23.0.0
hiredis==3.1.0
idna==3.10
packaging==24.2
pillow==11.1.0
powerline-status==2.7
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
python-dotenv==1.0.1
redis==5.2.1
reportlab==4.1.0
requests==2.32.3
setuptools==75.8.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
urllib3==2.3.0
weasyprint==64.0
webencodings==0.5.1
wheel==0.45.1
whitenoise==6.9.0
zopfli==0.2.3.post1
supabase-py==2.3.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
django-cache-memoize==0.2.0
django-debug-toolbar==4.3.0
sentry-sdk==1.40.4
django-storages==1.14.2
python-decouple==3.8
django-cacheops==7.0.2
